+++ example/lib/main_zh.dart
@@ -66,7 +66,7 @@ class _ExampleWidgetState extends State<ExampleWidget> {
                   height: ScreenUtil().setHeight(200),
                   color: Colors.red,
                   child: Text(
-                    '我的宽度:${ScreenUtil().setWidth(375)}dp \n'
+                    '我的宽度:${0.5.wp}dp \n'
                     '我的高度:${ScreenUtil().setHeight(200)}dp',
                     style: TextStyle(
                         color: Colors.white, fontSize: ScreenUtil().setSp(24)),
@@ -149,5 +149,8 @@ class _ExampleWidgetState extends State<ExampleWidget> {
         '宽度和字体相对于设计稿放大的比例:${ScreenUtil().scaleWidth * ScreenUtil.pixelRatio}');
     print('高度相对于设计稿放大的比例:${ScreenUtil().scaleHeight * ScreenUtil.pixelRatio}');
     print('系统的字体缩放比例:${ScreenUtil.textScaleFactor}');
+
+    print('屏幕宽度的0.5:${0.5.wp}');
+    print('屏幕高度的0.5:${0.5.hp}');
   }
 }